# CCIP-2971 Optimize token/gas prices database interactions