--- a/.changeset/thick-mails-applaud.md
+++ b/.changeset/thick-mails-applaud.md
@@ -1,0 +1,5 @@
+---
+"chainlink": patch
+---
+
+Simplify how token and gas prices are stored in the database - user upsert instead of insert/delete flow #db_update