Reference https://smartcontract-it.atlassian.net/browse/CCIP-2971


### Problem

Leader lane implementation inserts every price of every token/gas to the `observed_token_prices` and `observed_gas_prices` tables. It also evicts stale elements by removing them from the database, so all background workers (per lane) keep inserting token prices and deleting stale ones. Then leader lane worker is responsible for querying that table to pick only the relevant data (latest update per token)

It makes the entire process very db-heavy and inefficient because with every round we keep inserting new prices and evicting stale ones, which leads to:
* high dead tuples ratio - it impacts the storage by increasing IOPSes and pg overall performance by making auto vacuum busy
* high number of tuples fetched from that table - we scan all items from the index to pick the latest one by token_addr (it’s even higher than the evm.logs !)

### Solution

1. Get rid of deleting logic (simplify the code) and use upsert instead of insert. This will keep the table rather small - as many rows as tokens per chain. The table doesn’t grow in time so we don’t need the entire deletion logic - simplify the codebase, and reduce database connection usage by node. Token price updates only update the row in the database instead of creating a new row with every update. The huge benefit is also lightweight lookup (searching token prices by leader lane), which will be extremely small because the table is small.

2. Don’t insert if it’s not necessary. Multiple jobs might have the same tokens, before fetching the token price we can check which tokens didn’t get an update recently (to be defined what “recently” means in that case) and update only the ones that require it. Therefore we don’t flood the table with upsert requests.


### Tests

#### Go bench results 

Keep inserting random prices to the table. Please see `Benchmark_UpsertsTheSameTokenPrices`
```
// Before changes (always insert new logs). It's underestimated because it doesn't include deleting overhead 
Benchmark_UpsertsTheSameTokenPrices
Benchmark_UpsertsTheSameTokenPrices-12    	    2418	    483486 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    2662	    487156 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    2427	    514165 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    2481	    526274 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    2529	    507806 ns/op

// Always upsert
Benchmark_UpsertsTheSameTokenPrices
Benchmark_UpsertsTheSameTokenPrices-12    	    1047	   5467925 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    1204	   6483758 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	     897	   4814489 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    1300	   7206590 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    1112	   6422288 ns/op

// Upsert only relevant data (more read queries, less writes)
Benchmark_UpsertsTheSameTokenPrices
Benchmark_UpsertsTheSameTokenPrices-12    	    5346	    246378 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    4807	    245983 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    4275	    239922 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    4386	    248482 ns/op
Benchmark_UpsertsTheSameTokenPrices-12    	    4514	    254113 ns/op
```

#### CCIP baseline test

* Tuples fetches/read from index dropped from 1k to 400
* Dead tuples dropped from 4k to 300
* Similar Postgres CPU utilization 
* Higher index usage, from 3 RPS to 18 RPS. It's caused by:
   * we use an index when inserting to resolve conflicting entries
   * we fetch before upsert to reduce unnecessary updates 
* Although we scan the index more often, we still fetch less data than we used to fetch in the previous implementation  

Before

<img width="2539" alt="Screenshot 2024-08-14 at 14 16 00" src="https://github.com/user-attachments/assets/2d09e9d0-dd6c-404f-b2b1-cf7a7a15eae9">

After
<img width="2540" alt="Screenshot 2024-08-14 at 18 41 46" src="https://github.com/user-attachments/assets/ec75dff9-cb4d-4870-89c8-a3f0d18b53f4">