@mateusz-sekara 2024-08-19
@matYang should we backport the distinction between token and gas price updates here?
```go
	gasPriceUpdateInterval = 1 * time.Minute
	tokenPriceUpdateInterval = 10 * time.Minute
```