@matYang 2024-08-20
> @matYang should we backport from ccip repo the distinction between token and gas price updates here?
> 
> ```go
> 	gasPriceUpdateInterval = 1 * time.Minute
> 	tokenPriceUpdateInterval = 10 * time.Minute
> ```

yeh would be ideal, alternatively we can ignore the `PriceServices` changes here and only apply it inside ccip repo after merging chainlink to ccip, the latter should be easier