@mateusz-sekara 2024-08-20
> > @matYang should we backport from ccip repo the distinction between token and gas price updates here?
> > ```go
> > 	gasPriceUpdateInterval = 1 * time.Minute
> > 	tokenPriceUpdateInterval = 10 * time.Minute
> > ```
> 
> yeh would be ideal, alternatively we can ignore the `PriceServices` changes here and only apply it inside ccip repo after merging chainlink to ccip, btw there is another PR in ccip that touches PriceServices: https://github.com/smartcontractkit/ccip/pull/1275/files, from a merge perspective, the latter should be easier, do we wish to keep ccip code in chainlink repo somewhat up to date?

Ouch, that's the problem here. We need a single source of truth and unfortunately, PRs touching db migrations have to be done in chainlink repo first. Not sure what's the best approach here, @asoliman92 any suggestions?