@asoliman92 2024-08-20
>Ouch, that's the problem here. We need a single source of truth and unfortunately, PRs touching db migrations have to be done in chainlink repo first. Not sure what's the best approach here, @asoliman92 any suggestions?
We can either cherry-pick ccip changes to chainlink and resolve conflicts in this PR or merge that PR and resolve conflicts on ccip level

I've done the cherry-pick approach while doing the initial porting from CCIP to CL. It worked well at the time and I think it should be okay with your PR. 

I'm actually more interested in merging CL to CCIP and fixing the conflicts as we'll need to do this someway or another once V1.5 is ready. But this might be outside the scope of your PR.